% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean up earthquake data}
\usage{
eq_clean_data(raw_data, column_id)
}
\arguments{
\item{raw_data}{the data frame to be cleaned}

\item{column_id}{the column containting location  as a string}
}
\value{
a cleaned data frame
}
\description{
Cleans up data frame by creating a date column, converting lat/long, deaths, and magnitude to numeric rather than character, and applying eq_location_clean.  Function assumes data splits
date into YEAR, MONTH, and DAY column.  Location data needs to be in LATITUDE and LONGITUDE columns.
Function is case sensitive.  Future work should focus on making eq_clean_data operate like dplyr, taking column names rather than strings.
}
\examples{
\dontrun{
eq_clean_data(earthquakes, "LOCATION_NAME")
}

}
