% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create label for mapping earthquake data}
\usage{
eq_create_label(clean_data_frame)
}
\arguments{
\item{clean_data_frame}{Data frame of earthquake data that has been processed with eq_clean_data.}
}
\value{
List of character objects that can be used to create labels in eq_map.
}
\description{
Function creates a label column ("popup_text") in a data frame that has been cleaned with eq_clean_data.  The column contains html formatted text for use as a label
with the eq_map function. The label contains location, maginitude and deaths of each earthquake in the data frame.  If any earthquake is missing one of those things,
the label omits that line.
}
\examples{
data("earthquakes")
earthquakes \%>\% eq_clean_data("LOCATION_NAME") \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
eq_map(annot_col = "popup_text")

}
