% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{eq_map}
\alias{eq_map}
\title{Create earthquake map}
\usage{
eq_map(clean_data_frame, annot_col)
}
\arguments{
\item{clean_data_frame}{Data frame cleaned with eq_clean_data}

\item{annot_col}{Column to be used for creating label}
}
\value{
Leaflet map (technically returns NULL and draws map to plotting object)
}
\description{
Function takes a data frame cleaned with eq_clean_data and creates a leadlet map that plots earthquakes with size based on the magnitude of the earthquake.
The function also takes a column name (as a string) to use to create hoverover labels for the point. Future work should focus on making eq_clean_data operate
like dplyr, taking column names rather than strings.
}
\examples{
\dontrun{
earthquakes \%>\% eq_clean_data("LOCATION_NAME") \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
eq_map(annot_col = "DATE")
}

}
